f4pga: merge sfbuild into __init__
